+++ src/main/java/com/example/vkrfull/security/SignupRequest.java
@@ -17,6 +17,6 @@ public class SignupRequest {
 
     private String password;
 
-    private Long student;
+    private Integer student;
 
 }
